@@ -14,13 +14,13 @@
         }
     ],
     "require": {
-        "php": ">=7.4.0",
-        "psr/log": "~1.0",
+        "php": ">=8.1.0",
+        "psr/log": "^3.0",
         "xenos/asklucy": "^2"
     },
     "require-dev": {
         "guzzlehttp/guzzle": "^6",
-        "phpunit/phpunit": "^6.2",
+        "phpunit/phpunit": "10@dev",
         "squizlabs/php_codesniffer": "*"
     },
     "suggest": {
