@@ -16,7 +16,7 @@ trait CountableTrait
      *
      * @return int
      */
-    public function count()
+    public function count(): int
     {
         return count($this->elements);
     }
