@@ -15,7 +15,7 @@ use MusicBrainz\Value\Name;
  * @see  RemixesAndCompilations
  * @link https://musicbrainz.org/relationship/6cc958c0-533b-4540-a281-058fbb941890
  */
-class Mixed extends Engineered
+class AudioMixed extends Engineered
 {
     /**
      * Returns the name of the relation.
